requirements.txt
django
cleverbot.py
